+++ src/test/java/com/mindee/v2/product/ExtractionTest.java
@@ -11,7 +11,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.mindee.geometry.Point;
 import com.mindee.geometry.Polygon;
-import com.mindee.v2.input.LocalResponse;
+import com.mindee.v2.parsing.LocalResponse;
 import com.mindee.v2.parsing.inference.InferenceActiveOptions;
 import com.mindee.v2.parsing.inference.InferenceFile;
 import com.mindee.v2.parsing.inference.InferenceJob;
